+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Other changes
 
   * `stubbles\sequence\Sequence` can now be casted to string, which provides information about how the sequence is build
+  * added `stubbles\sequence\assert\provides()` which creates a predicate that allows to assert that a sequence contains the expected values
 
 
 7.0.0 (2016-01-11)